Add dynamic errors for SQS
Dynamic errors are not present in the JSON file, but may be returned as an API response.
Currently, there is no way to get a list of these errors other than from the documentation for each service.
In aws-sdk-ruby, these errors are dynamically generated and can be specified with `rescue`,
but since the type does not exist, they are considered a non-existent error class.
As an experiment, we will try to generate the type of this dynamic error class based on the documentation.