activerecord: Add modules for define a subclass of ActiveRecord::Relation
Add new modules `ActiveRecord::Relation::Methods` and `ActiveRecord::
Relation::ClassMethods` to define a subclass of ActiveRecord::Relation.

So far, we have provided `_ActiveRecord_Relation` interface and
`_ActiveRecord_Relation_ClassMethods` interface for the same purpose.
However, using interfaces does not support enhancement.  For example,
users can't override the method types on the subclasses.

Therefore this starts providing modules to support the enhancement of the
subclasses.  The methods of them are the same as what the interface
provides.

refs: https://github.com/pocke/rbs_rails/pull/254#issuecomment-1600115112

Note: Old interfaces should be keep as is for a while to keep
compatibility.