Add RBS files for sinatra and sinatra-contrib gems
Generated initial RBS files using the `rbs prototype rb` command,
and added type definitions for both the Sinatra core and contributed extensions.

https://github.com/sinatra/sinatra/tree/v4.0.0

I have set myself as the reviewer for this change because I believe
the sinatra gem is an important and widely-used web framework.
However, I am not very familiar with sinatra, so if there is someone who
uses sinatra regularly and is motivated to maintain these type definitions,
I would appreciate it if they could take over the reviewer role.