Using [self: instance] annotation in sinatra gem (#569)
When a method is called within a block of HTTP methods such as `get`, `post` in the sinatra gem,
it should have been treated as calling an instance method, but was evaluated as a singleton method.
By adding `[self: instance]` to the method block of each HTTP method,
processing within the block is now treated as processing in the instance.