activemodel: Remove ActiveModel::Attributes#initialize (#641)
Now `ActiveModel::Attributes#initialize` has been typed as `() -> untyped`.
But it's incorrect. The original implementation takes positional
arguments via varargs and passes it to other constructors using `super`
method. So it does not care the number of arguments.

refs:

* https://github.com/rails/rails/blob/v6.1.7.8/activemodel/lib/active_model/attributes.rb#L75
* https://github.com/rails/rails/blob/v7.2.1/activemodel/lib/active_model/attributes.rb#L106

Usually, many models using ActiveModel::Model takes keyword arguments
(ex. `Person.new(name: 'John')`. But, unfortunately, the current type of
ActiveModel::Atributes conceals the type of constructors for them.

This removes the definition of ActiveModel::Attributes#initialize
not to conceal the type of constructor.