Remove manifest.yaml from sidekiq-pro
sidekiq-pro declares a dependency to `sidekiq` gem with `manifest.yaml`.
It's because sidekiq-pro is a private gem, so it can't fetch sidekiq gem
in the test of this type definition.

But this `manifest.yaml` introduces an error when using this RBS via
`rbs collection`.

`rbs validate` command displays the following error.

```
Cannot find type definitions for library: sidekiq (0) (RBS::EnvironmentLoader::UnknownLibraryError)
```

`rbs collection install` generates the following lock file.

```
- name: sidekiq
  version: '0'
  source:
    type: stdlib
```

Because `manifest.yaml` works only for standard libraries. So
`manifest.yaml` does not work for a ordinary gem (at least for now).
See also https://github.com/ruby/rbs/issues/1538

So I removed this `manifest.yaml`.

----

BTW, this `manifest.yaml` is unnecessary for users.
The comment in the `manifest.yaml` said:

> manifest.yaml describes dependencies which do not appear in the gemspec.

It's for gems "which do not appear in the gemspec", but `sidekiq-pro`
gem declares a dependency to `sidekiq` in the gemspec. So the
`manifest.yaml` is not necessary the user to use this RBS definitions.