activesupport: Update return types of `#present?` and `#blank?`
This updates the return types of `#present?` and `#blank?` to narrow the
type of receiver possibly.

Since 1.10, Steep can narrow the type of receiver via the return type of
the method when the type of receiver is union.  For example, a receiver
typed as will be narrowed to `String` on the `if present?` block, and
be narrowed to `String | nil` on the `else` block because
`String#present?` returns bool and `NilClass#present?` returns false.

```ruby
var = ... # String | nil
if var.present?
  # var is narrowed to String
else
  # var is narrowed to String | nil
end
```

refs:

* https://github.com/soutaro/steep/wiki/Release-Note-1.10 (unreleased yet)
* https://github.com/soutaro/steep/pull/1497