fix activerecord sum methods (#814)
When a block is given, it is delegated to Enumerable#sum. See `Enumerable#sum` for reference.

Since it is more type-friendly for the types of `initial_value`, the block's return value, and the method's return value to match, I will accept some imprecision and unify all types as `T`.

When no block is given, the calculate method processes the result, so the return value is not necessarily an Integer.

https://github.com/rails/rails/blob/v6.0.6.1/activerecord/lib/active_record/relation/calculations.rb#L127